Добавлены необходимые права в аргументах команды

+++ src/main/java/commands/SayCommand.java
@@ -3,7 +3,7 @@ package commands;
 import api.BasicEmbed;
 import api.models.command.Command;
 import api.models.command.DiscordCommand;
-import api.models.workers.WorkerHandler;
+import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 
 @DiscordCommand(name = "say", aliases = {"сказать", "скажи"}, arguments = 1, group = "Утилиты")
@@ -11,9 +11,10 @@ public class SayCommand implements Command {
 
     @Override
     public void doCommand(MessageReceivedEvent msg_event, String[] arguments) {
-        System.out.println(WorkerHandler.workersThreads);
         if (arguments.length != 0) {
-            msg_event.getMessage().delete().queue();
+            if (msg_event.getGuild().getSelfMember().hasPermission(Permission.MESSAGE_MANAGE)) {
+                msg_event.getMessage().delete().queue();
+            }
             msg_event.getChannel().sendMessage(String.join(" ", arguments)).queue();
         } else {
             BasicEmbed errorEmbed = new BasicEmbed("error");
